redpanda_to_postgres/requirements.txt
kafka-python-ng
confluent-kafka 
psycopg2-binary